fix(flux-lsp): split binary from library
in order to compile to wasm the bulk of the
code needs to be in the flux_lsp_lib
